Move fake testing AMQP server into its own file
